Fix compile errors.
