+++ test/support/src/vfs_helpers.cc
@@ -45,6 +45,10 @@
 #include "test/support/src/helpers.h"
 #include "test/support/src/vfs_helpers.h"
 
+#if HAVE_S3
+#include "tiledb/sm/filesystem/s3.h"
+#endif
+
 #include "tiledb/sm/rest/rest_client.h"
 
 namespace tiledb::test {
@@ -494,6 +498,27 @@ VFSTest::VFSTest(
     }
   }
   std::sort(expected_results().begin(), expected_results().end());
+}
+
+S3Test::S3Test(const std::vector<size_t>& test_tree)
+    : VFSTestBase(test_tree, "s3://")
+    , S3_within_VFS(&tiledb::test::g_helper_stats, &io_, vfs_.config()) {
+#ifdef HAVE_S3
+  s3().create_bucket(temp_dir_);
+  for (size_t i = 1; i <= test_tree_.size(); i++) {
+    sm::URI path = temp_dir_.join_path("subdir_" + std::to_string(i));
+    // VFS::create_dir is a no-op for S3; Just create objects.
+    for (size_t j = 1; j <= test_tree_[i - 1]; j++) {
+      auto object_uri = path.join_path("test_file_" + std::to_string(j));
+      s3().touch(object_uri);
+      std::string data(j * 10, 'a');
+      s3().write(object_uri, data.data(), data.size());
+      s3().flush(object_uri);
+      expected_results().emplace_back(object_uri.to_string(), data.size());
+    }
+  }
+  std::sort(expected_results().begin(), expected_results().end());
+#endif
 }
 
 LocalFsTest::LocalFsTest(const std::vector<size_t>& test_tree)
